+++ src/test/pegtl/actions_one.cpp
@@ -29,16 +29,16 @@ namespace TAO_PEGTL_NAMESPACE
       {
          TAO_PEGTL_TEST_ASSERT( applied.size() == 10 );
 
-         TAO_PEGTL_TEST_ASSERT( applied[ 0 ].first == tao::demangle< one< 'b' > >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 1 ].first == tao::demangle< foo >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 2 ].first == tao::demangle< at< one< 'a' > > >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 3 ].first == tao::demangle< two< 'a' > >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 4 ].first == tao::demangle< fiz >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 5 ].first == tao::demangle< foo >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 6 ].first == tao::demangle< one< 'b' > >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 7 ].first == tao::demangle< foo >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 8 ].first == tao::demangle< eof >() );
-         TAO_PEGTL_TEST_ASSERT( applied[ 9 ].first == tao::demangle< bar >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 0 ].first == demangle< one< 'b' > >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 1 ].first == demangle< foo >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 2 ].first == demangle< at< one< 'a' > > >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 3 ].first == demangle< two< 'a' > >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 4 ].first == demangle< fiz >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 5 ].first == demangle< foo >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 6 ].first == demangle< one< 'b' > >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 7 ].first == demangle< foo >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 8 ].first == demangle< eof >() );
+         TAO_PEGTL_TEST_ASSERT( applied[ 9 ].first == demangle< bar >() );
 
          TAO_PEGTL_TEST_ASSERT( applied[ 0 ].second == "b" );
          TAO_PEGTL_TEST_ASSERT( applied[ 1 ].second == "b" );
@@ -60,7 +60,7 @@ namespace TAO_PEGTL_NAMESPACE
       template< typename ActionInput >
       static void apply( const ActionInput& in )
       {
-         applied.emplace_back( tao::demangle< Rule >(), in.string() );
+         applied.emplace_back( demangle< Rule >(), in.string() );
       }
    };
 
@@ -69,7 +69,7 @@ namespace TAO_PEGTL_NAMESPACE
       parse< disable< test1::bar >, test_action >( memory_input( "baab", __FUNCTION__ ) );
       TAO_PEGTL_TEST_ASSERT( applied.size() == 1 );
 
-      TAO_PEGTL_TEST_ASSERT( applied[ 0 ].first == tao::demangle< disable< test1::bar > >() );
+      TAO_PEGTL_TEST_ASSERT( applied[ 0 ].first == demangle< disable< test1::bar > >() );
       TAO_PEGTL_TEST_ASSERT( applied[ 0 ].second == "baab" );
 
       applied.clear();